Make Illusts model-based
Doing it the previous way (QML list property) is... fine but it should
really be avoided. It creates bindings which are slow to update -
especially for properties this big - and if we want to say, add more
illustrations to the list then QML has to re-create the entire list
which makes the application much slower.

The previous illusts API was not removed yet, so it will continue to
function.